Create lib/entities/idol.lua Move idol code into idol.lua
